Port to Qt 6
Qt 6 treats high DPI differently than Qt 5,
resulting in pixellated widgets on all operating systems
(whereas Qt 5 looked fine on Windows but pixellated on Linux).
This makes it more urgent to fix the appearance on all OSes.

+++ src/gui/lib/color.h
@@ -4,15 +4,15 @@
 
 namespace gui::lib::color {
 
-inline qreal lerp(qreal x, qreal y, qreal position) {
+inline float lerp(float x, float y, float position) {
     return x + position * (y - x);
 }
 
 /// Blends two colors in linear color space.
 /// Produces better results on both light and dark themes,
 /// than integer blending (which is too dark).
-QColor lerp_colors(QColor c1, QColor c2, qreal position);
+QColor lerp_colors(QColor c1, QColor c2, float position);
 
-QColor lerp_srgb(QColor c1, QColor c2, qreal position);
+QColor lerp_srgb(QColor c1, QColor c2, float position);
 
 }
